DX-2965: let a host CLI mount the box command tree on an OAuth login
The SDK sends a JWT credential as Authorization: Bearer, which is where
box-coordinator reads OAuth access tokens; Box API keys keep X-Box-Api-Key.
The CLI's command tree moves into buildBoxProgram() so another CLI can add it
as a subcommand, and setDefaultToken() lets that host supply the token behind
--token and UPSTASH_BOX_API_KEY.
