+++ .changeset/dx-2965-box-oauth-cli.md
@@ -1,0 +1,5 @@
+---
+"@upstash/box-cli": minor
+---
+
+Export `buildBoxProgram` and `setDefaultToken` so a host CLI can mount the `box` command tree as a subcommand and supply the token from its own login.
