+++ .changeset/dx-2965-box-oauth-sdk.md
@@ -1,0 +1,5 @@
+---
+"@upstash/box": patch
+---
+
+Send a JWT credential (an OAuth access token from an Upstash login) as `Authorization: Bearer`; Box API keys keep using `X-Box-Api-Key`.
